Move around directories and files
Refs #41.
