+++ devices/chopper/device.py
@@ -18,8 +18,10 @@
 # *********************************************************************
 
 from collections import OrderedDict
-from simulation.core import StateMachine, CanProcessComposite, CanProcess, Context
-from simulation.core.utils import dict_strict_update
+
+from core import StateMachine, CanProcessComposite, CanProcess, Context
+from core.utils import dict_strict_update
+
 from bearings import MagneticBearings
 from defaults import *
