+++ setups/__init__.py
@@ -18,15 +18,16 @@
 # *********************************************************************
 
 import importlib
-from simulation.core import CanProcess
+
+from core import CanProcess
 
 
 def import_device(device_type, scenario):
     """
-    This function is a helper that imports the first object which is an instance of simulation.core.CanProcess
-    from the scenarios package:
+    This function is a helper that imports the first object which is an instance of devices.core.CanProcess
+    from the setups package:
 
-        from scenarios.device_type.scenario import can_process_object.
+        from setups.device_type.scenario import can_process_object.
 
     The object is returned by the function, so to import the default scenario for chopper:
 
@@ -37,7 +38,7 @@ def import_device(device_type, scenario):
     :return: Device object as specified by device_type and scenario
     """
     module_name = '.{}'.format(scenario)
-    scenario_package = 'scenarios.{}'.format(device_type)
+    scenario_package = 'setups.{}'.format(device_type)
 
     module = importlib.import_module(module_name, scenario_package)
 
@@ -53,7 +54,7 @@ def import_device(device_type, scenario):
 
 def import_bindings(device_type, bindings_type):
     """
-    This function imports a variable named bindings_type from scenarios.device_type.bindings.
+    This function imports a variable named bindings_type from setups.device_type.bindings.
     This relies on the convention that the name of the bindings-variable has the same name as the
     communication protocol it addresses. For example, importing the EPICS bindings for simulated choppers:
 
@@ -61,12 +62,12 @@ def import_bindings(device_type, bindings_type):
 
     This is equivalent to:
 
-        from scenarios.chopper.bindings import epics as bindings
+        from setups.chopper.bindings import epics as bindings
 
     :param device_type: Device for which the bindings are written
     :param bindings_type: Name of the variable to import as bindings
     :return: Variable that specifies the binding.
     """
-    module = importlib.import_module('.bindings', 'scenarios.{}'.format(device_type))
+    module = importlib.import_module('.bindings', 'setups.{}'.format(device_type))
 
     return getattr(module, bindings_type)
